Allow Cards To Be Forced Into Deck
Allows cards to be forced into a deck by passing a "True" value after
the IsMerit value. Allows forcing of a specific deck this way with all
"True" values. Even if a full deck is forced, the deck will be fully
shuffled. This feature does not interfere with previous functionality
and is fully backwards-compatible. Also calculates the player deck cost
once instead of once for each deck combination (small optimization).